Fix up more tests
