+++ tests/test_query.py
@@ -1,14 +1,25 @@
 import pytest
 
 from graphene.test import Client
-from schema import schema
+
+from example.database import load_from_raw
+from example.schema import schema
 
 # test example
 # see https://docs.graphene-python.org/en/latest/testing/
 
-def test_all_houses():
+
+def add_test_data(session):
+	house_models, student_models, staff_models = load_from_raw()
+	for model in house_models + student_models + staff_models:
+		session.add(model)
+	session.commit()
+
+
+def test_all_houses(session):
+	add_test_data(session)
 	client = Client(schema)
-	query = '''
+	query = """
 	{
 	  allHouses {
 	    edges {
@@ -17,8 +28,8 @@ def test_all_houses():
 	      }
 	    }
 	  }
-	}'''
-	result = {
+	}"""
+	expected_result = {
 	  "data": {
 	    "allHouses": {
 	      "edges": [
@@ -48,4 +59,5 @@ def test_all_houses():
 	}
 
 	executed = client.execute(query)
-	assert executed == result
+	print(executed)
+	assert executed == expected_result
